feat: (src/main) smallest no_std && no_main source
The OS entry cannot need `std` and `main` because those need another OS
to suppost.

Instead of `main`, there is a function named `_start`, which is the
entry point of this project. Because the linker looks for this function
named `_start` by default.

To deal with panic, there is a function `panic` to try to deal with it.

To deal with abort, there is a function `abort` to try to deal with it.

And it also need an language item, so there is a `eh_personality` here.

for more infomation, click:
    https://intermezzos.github.io/book/first-edition/creating-our-first-crate.html
    (zh) https://learningos.github.io/rcore_step_by_step_webdoc/docs/%E7%8B%AC%E7%AB%8B%E5%BC%8F%E5%8F%AF%E6%89%A7%E8%A1%8C%E7%A8%8B%E5%BA%8F.html