feat: (src/interrupt.rs, *) add funtion to handle interrupt
Handling interrupt is important for a OS-Kernel.

To handle it, I use `crate::interrupt::init()` in rust_main in
`src/init.rs` to set the address of `__alltraps` to register `stvec` to
try to handle the interrupt.

The function `__alltraps` will save all register it need to a struct
`TrapFrame` to save the context, which is defined in `src/context.rs`.

I also change the `Cargo.toml` to use `riscv` to handle its registers.